refactor: 섹션을 폴더로 분리
